requirements.txt
（原始版本）
torch>=1.8.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0